+++ changelog.md
@@ -1,6 +1,7 @@
 ### 0.0.2 [Feb 2023]
 - **[new]** - number parsing
 - **[new]** - verb conjugation
+- **[new]** - tagger - 80% accuracy
 
 ### 0.0.1 [Feb 2023]
 - **[new]** - tokenization